requirement.txt.txt
tensorflow
streamlit
opencv-python-headless
numpy
Pillow
gdown